chore: rename UniversalBitcoinAmountInput -> BitcoinAmountInput
