+++ src/components/Send/SendForm.tsx
@@ -9,7 +9,7 @@ import { SourceJarSelector } from './SourceJarSelector'
 import { CoinjoinPreconditionViolationAlert } from '../CoinjoinPreconditionViolationAlert'
 import { DestinationInputField, DestinationValue } from './DestinationInputField'
 import { AmountInputField } from './AmountInputField'
-import { AmountValue } from '../UniversalBitcoinAmountInput'
+import { AmountValue } from '../BitcoinAmountInput'
 import CollaboratorsSelector from './CollaboratorsSelector'
 import { SweepBreakdown } from './SweepBreakdown'
 import FeeBreakdown from './FeeBreakdown'
